Harden ensure_local_checks against varied environs
To ensure that each appropriate check can be applied in an
environment which contains affinity > 1 of any container we
will append the inventory_hostname to the check's yaml file
so that we can have multiplicity of checks per host.

Also removed unnecessary commented out lines at the end of the file.