Add alarm verification step to maas role
For every check, ensure that at least one alarm is attached to it.

This involves two primary changes:

- Add the verify tasks for alarms and invoke it after checks are
verified.

- Modify the tab format for checks and alarms to include the filename
the check originated from, if any. This simplifies the matching of
check labels listed from API against those on the filesystem and
eliminates some of the complexity with respect to needing to keep
these in sync resulting in a simpler system over all.

The term verify is used instead of validate as a secondary change.