RE-1287 Allow openstack_deploy as symlink
Currently get-maas.yml checks that /etc/openstack_deploy is a directory,
whereas the check was intended to ensure that it exists (according to
the naming of the task.)

This patch uses the stat and fail modules to ensure that
openstack_deploy exists, without forcing it to be a particular type.

This is useful when you want to keep config in git, and symlink a dir
from that repo to /etc/openstack_deploy.