(ASC-219) Manage exit failures with trap
This commit alters the `run_system_tests.sh` script to manage failures
with a custom `trap`. This enables all steps of the script to get
executed, but will return the exit code of the first failing command.

The setup portion of the script is set to fail immediately, while the
script execution and artifact gathering portions will all get executed
in order to gather any log and test results that may be available. The
failing exit code of any process is captured in a global `RC` variable
and is returned as the scripts exit code.

This ensures the Jenkins job will be marked as failed if the test step
or one of the artifact gathering steps fail.